+++ GoTrue/src/desktopMain/kotlin/io/github/jan/supabase/gotrue/server/HttpCallbackHtml.kt
@@ -1,0 +1,52 @@
+package io.github.jan.supabase.gotrue.server
+
+internal object HttpCallbackHtml {
+
+    fun landingPage(title: String) = """
+            <!DOCTYPE html>
+            <html lang="en">
+                  <head>
+                        <title>$title</title>
+                  </head>
+
+                  <body>
+                        <script>
+                            const pairs = location.hash.substring(1).split("&").map(pair => pair.split("="))
+                            const accessToken = pairs.find(pair => pair[0] === "access_token")[1]
+                            const refreshToken = pairs.find(pair => pair[0] === "refresh_token")[1]
+                            const expiresIn = pairs.find(pair => pair[0] === "expires_in")[1]
+                            const tokenType = pairs.find(pair => pair[0] === "token_type")[1]
+                            location.href = "/callback?access_token=" + accessToken + "&refresh_token=" + refreshToken + "&expires_in=" + expiresIn + "&token_type=" + tokenType
+                       </script>
+                 </body>
+            </html>
+        """.trimIndent()
+
+    fun redirectPage(icon: String, title: String, text: String) = """
+        <!DOCTYPE html>
+        <html>
+            <head>
+                <meta charset="utf-8">
+                <link rel="icon" href="$icon">
+                <title>$title</title>
+            </head>
+            <body style="background-color:#2f3237;">
+                <p style="    
+                    position: absolute;
+                    top: 50%;
+                    font-size: 2.5em;
+                    left: 50%;
+                    -moz-transform: translateX(-50%) translateY(-50%);
+                    -webkit-transform: translateX(-50%) translateY(-50%);
+                    transform: translateX(-50%) translateY(-50%);
+                    font-family: Uni Sans, sans-serif;
+                ">$text</p>
+                <script>
+                    const newURL = location.href.split("?")[0];
+                    window.history.replaceState({}, document.title, newURL);
+                </script>
+            </body>
+        </html>
+    """.trimIndent()
+
+}
